Fixed problems regarding comments
