+++ llvm/lib/Transforms/PeepholeOptimizationCourse/Obfuscate8Bytes/Obfuscate8BytesAdd.cpp
@@ -1,3 +1,5 @@
+#include "llvm/Transforms/PeepholeOptimizationCourse/Obfuscate8BytesNewPassManager.h"
+
 #include "llvm/Pass.h"
 #include "llvm/IR/Function.h"
 #include "llvm/IR/InstIterator.h"
@@ -14,6 +16,30 @@
 using namespace llvm;
 namespace {
 
+	bool isAddWithI8(Instruction const* instruction) {
+		bool is8Bit = operand->getType()->isIntegerTy() && (operand->getType()->getIntegerBitWidth() == 8);
+		bool isBinaryAdd = (instruction->getOpcode() == Instruction::Add) && (instruction->getNumOperands() == 2);
+		return is8Bit && isBinaryAdd;
+	}
+
+	void doObfuscateI8Add(Instruction* instruction) {
+		auto const& a = instruction->getOperand(0);
+		auto const& b = instruction->getOperand(1);
+		auto const& I8Type = a->getType();
+		
+		IRBuilder<> Builder(instruction);
+		auto mul4 =  Builder.CreateXor(a,b);
+		auto inc2 = Builder.CreateAnd(a, b);
+		auto inc3 = Builder.CreateMul(ConstantInt::get(I8Type, 2), mul4);
+		auto mul3 = Builder.CreateAdd(inc2, inc3);
+		auto mul2 = Builder.CreateMul(ConstantInt::get(I8Type, 39), mul3);
+		auto mul1 = Builder.CreateAdd(mul2, ConstantInt::get(I8Type, 23));
+		auto inc1 = Builder.CreateMul(mul1, ConstantInt::get(I8Type, 151));
+		Instruction *newInst = BinaryOperator::CreateAdd(inc1, ConstantInt::get(I8Type, 111));
+		ReplaceInstWithInst(instruction, newInst);	
+	}
+
+
 	struct Obfuscate8Bytes : public FunctionPass {
     static char ID;
     Obfuscate8Bytes() : FunctionPass(ID) {}
@@ -27,42 +53,11 @@ namespace {
         for (auto basicBlock = function.begin(), basicBlockEnd = function.end(); basicBlock != basicBlockEnd; ++basicBlock) {
             for (auto instruction = basicBlock->begin(), ie = basicBlock->end(); instruction != ie; ++instruction) {
                 auto binaryOperator = dyn_cast<BinaryOperator>(instruction);
-                if (!binaryOperator) {
-                    continue;
-                }
-
-                if (binaryOperator->getOpcode() != Instruction::Add) {
-                    continue;
-                }
-
-                if (!binaryOperator->getType()->isIntegerTy() || binaryOperator->getType()->getIntegerBitWidth() != 8) {
-                    continue;
-                }
-
-                auto Const39 = ConstantInt::get(binaryOperator->getType(), 39);
-                auto Const151 = ConstantInt::get(binaryOperator->getType(), 151);
-                auto Const23 = ConstantInt::get(binaryOperator->getType(), 23);
-                auto Const2 = ConstantInt::get(binaryOperator->getType(), 2);
-                auto Const111 = ConstantInt::get(binaryOperator->getType(), 111);
-				// 111 + (151 * (23 + (39 * ((a |! b) + (2 * (a+b))))))
-                IRBuilder<> builder(binaryOperator);
-                auto obfuscated = BinaryOperator::CreateAdd(Const111,
-                    builder.CreateMul(Const151,
-                        builder.CreateAdd(Const23,
-                            builder.CreateMul(Const39,
-                                builder.CreateAdd(
-                                    builder.CreateXor(binaryOperator->getOperand(0), binaryOperator->getOperand(1)),
-                                    builder.CreateMul(Const2,
-                                    	builder.CreateAnd(binaryOperator->getOperand(0), binaryOperator->getOperand(1)
-                                    	)
-                                    )
-                                )
-                            )
-                        )
-                    )
-                );
-
-                ReplaceInstWithInst(basicBlock->getInstList(), instruction, obfuscated);
+                
+				if (!isAddWithI8(instruction))
+					continue;
+				
+				doObfuscateI8Add(instruction);
 
                 LLVM_DEBUG(dbgs() << *binaryOperator << " -> " << *obfuscated << '\n');
                 changed = true;
